# doc/api.yaml
openapi: 3.0.0
info:
  title: 'Home Library Service'
  description: 'Home music library service'
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Freddie Mercury'
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
...
paths:
...
  /artist:
    get:
      tags:
        - Artists
      summary: 'Get all artists'
      description: 'Gets all artists'
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
    post:
      tags:
        - Artists
      summary: 'Add a new artist'
      description: 'Adds a new artist'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Freddie Mercury'
                grammy:
                  type: boolean
                  example: false
              required:
                - name
                - grammy
      responses:
        201:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: 'Bad request. Body does not contain required fields or values are invalid'
  /artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Artists
      summary: 'Get a single artist by id'
      description: 'Gets a single artist by id'
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: 'Bad request. artistId is invalid (not uuid)'
        404:
          description: 'Artist was not found'
    put:
      tags:
        - Artists
      summary: 'Update artist information'
      description: 'Updates artist information by id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Bob Dylan'
                grammy:
                  type: boolean
                  example: false
              required:
                - name
                - grammy
      responses:
        200:
          description: 'The artist has been updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: 'Bad request. artistId is invalid (not uuid) or values are invalid'
        404:
          description: 'Artist was not found'
    delete:
      tags:
        - Artists
      summary: 'Delete an artist'
      description: 'Deletes an artist from library'
      responses:
        204:
          description: 'Delelted succesfully'
        400:
          description: 'Bad request. artistId is invalid (not uuid)'
        404:
          description: 'Artist was not found'
...
  /favs/artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: 'Add an artist to the favorites'
      description: 'Adds an artist to the favorites'
      responses:
        201:
          description: 'Artist has been added to favorites'
        400:
          description: 'Bad request. artistId is invalid (not uuid)'
        422:
          description: 'Artist with id does not exist'
    delete:
      tags:
        - Favorites
      summary: 'Delete an artist from favorites'
      description: 'Deletes an artist from favorites'
      responses:
        204:
          description: 'Deleted succesfully'
        400:
          description: 'Bad request. artistId is invalid (not uuid)'
        404:
          description: 'Artist was not found'